Added the CONFIG_OPENSSL flag to disable https support at build time
