+++ http.h
@@ -56,9 +56,13 @@ int parse_uri(const char *uri, struct http_uri *u);
 /* frees contents of @u, not @u itself */
 void http_free_uri(struct http_uri *u);
 
+int get_sockfd(struct connection *conn);
 int socket_open(struct http_get *hg, int timeout_ms);
 int connection_open(struct connection *conn, struct http_get *hg, int timeout_ms);
 int connection_close(struct connection *conn);
+
+int socket_write(struct connection *conn, const char *in_buf, int count);
+int socket_read(struct connection *conn, char *out_buf, int count);
 
 /*
  * returns:  0 success
